:shirt: address review comments
