+++ fs/base.py
@@ -8,7 +8,6 @@ can work with any of the supported filesystems.
 
 from __future__ import absolute_import, print_function, unicode_literals
 
-import re
 import abc
 import hashlib
 import itertools
@@ -16,7 +15,6 @@ import os
 import threading
 import time
 import typing
-import platform
 from contextlib import closing
 from functools import partial, wraps
 import warnings
@@ -63,7 +61,6 @@ if False:  # typing.TYPE_CHECKING
 
 
 __all__ = ["FS"]
-_WINDOWS_PLATFORM = platform.system() == "Windows"
 
 
 def _new_name(method, old_name):
@@ -312,20 +309,6 @@ class FS(object):
             >>> my_fs.setinfo('file.txt', details_info)
 
         """
-
-    @staticmethod
-    def quote(path_snippet):
-        if _WINDOWS_PLATFORM and has_drive_letter(path_snippet):
-            drive_letter, path = path_snippet.split(":", 1)
-            if six.PY2:
-                path = path.encode("utf-8")
-            path = six.moves.urllib.request.pathname2url(path)
-            path_snippet = "{}:{}".format(drive_letter, path)
-        else:
-            if six.PY2:
-                path_snippet = path_snippet.encode("utf-8")
-            path_snippet = six.moves.urllib.request.pathname2url(path_snippet)
-        return path_snippet
 
     # ---------------------------------------------------------------- #
     # Optional methods                                                 #
@@ -1662,8 +1645,3 @@ class FS(object):
                     break
                 hash_object.update(chunk)
         return hash_object.hexdigest()
-
-
-def has_drive_letter(path_snippet):
-    windows_drive_pattern = ".:[/\\\\].*$"
-    return re.match(windows_drive_pattern, path_snippet) is not None
